Package Update

+++ app/Http/Controllers/Landlord/CustomerController.php
@@ -144,21 +144,69 @@ class CustomerController extends Controller
         return response()->json($result['alertMsg'], $result['statusCode']);
     }
 
+    // public function changePackageProcess(Request $request, Tenant $tenant)
+    // {
+    //     DB::beginTransaction();
+    //     try {
+
+    //         $tenant->package_id = $request->package_id;
+    //         $tenant->update();
+
+    //         $permissions = json_decode($tenant->package->permissions, true);
+    //         $allPermissionIds = explode(',',$tenant->package->permission_ids);
+
+    //         $tenant->run(function ($tenant) use ($permissions, $allPermissionIds) {
+
+                // DB::table('permissions')->delete();
+                // DB::table('permissions')->insert($permissions);
+                // $role = Role::find(1);
+                // $role->syncPermissions($allPermissionIds);
+    //         });
+    //         $result =  Alert::successMessage('Package Switched Successfully');
+    //         DB::commit();
+
+    //     } catch (Exception $e) {
+    //         DB::rollback();
+    //         $result =  Alert::errorMessage($e->getMessage());
+    //     }
+
+    //     return response()->json($result['alertMsg'], $result['statusCode']);
+    // }
     public function changePackageProcess(Request $request, Tenant $tenant)
     {
         DB::beginTransaction();
         try {
+
+            $prevPermissions = json_decode($tenant->package->permissions, true);
+            $prevPermissionIds = array_column($prevPermissions, 'id');
+
             $tenant->package_id = $request->package_id;
             $tenant->update();
 
-            $permissions = json_decode($tenant->package->permissions, true);
-            $allPermissionIds = explode(',',$tenant->package->permission_ids);
+            $package = Package::find($request->package_id);
+            $latestPermissions = json_decode($package->permissions, true);
+            $latestPermissionsIds = array_column($latestPermissions, 'id');
 
-            $tenant->run(function ($tenant) use ($request, $permissions, $allPermissionIds) {
-                DB::table('permissions')->delete();
-                DB::table('permissions')->insert($permissions);
-                $role = Role::find(1);
-                $role->syncPermissions($allPermissionIds);
+            $newAddedPermissions = [];
+            foreach ($latestPermissions as $element) {
+                if (!in_array($element["id"], $prevPermissionIds)) {
+                    $newAddedPermissions[] = $element;
+                }
+            }
+
+            // $permissions = json_decode($tenant->package->permissions, true);
+            // $allPermissionIds = explode(',',$tenant->package->permission_ids);
+
+            $tenant->run(function ($tenant) use ($newAddedPermissions, $latestPermissionsIds) {
+                // DB::table('model_has_permissions')->truncate();
+                // DB::table('permissions')->truncate();
+                // DB::table('permissions')->delete();
+                // DB::table('permissions')->insert($latestPermissions);
+
+                DB::table('permissions')->whereNotIn('id', $latestPermissionsIds)->delete();
+                DB::table('permissions')->insert($newAddedPermissions);
+                $role = Role::findById(1);
+                $role->syncPermissions($latestPermissionsIds);
             });
             $result =  Alert::successMessage('Package Switched Successfully');
             DB::commit();
@@ -232,4 +280,7 @@ class CustomerController extends Controller
         // $tenant->delete();
         // return 'ok';
     }
+
+    // php artisan cache:forget spatie.permission.cache
+    // php artisan cache:clear
 }
